+++ src/OneBot/V12/Driver/WorkermanDriver.php
@@ -6,10 +6,10 @@ namespace OneBot\V12\Driver;
 
 use Error;
 use MessagePack\MessagePack;
+use OneBot\Util\MPUtils;
 use OneBot\V12\Action\ActionResponse;
 use OneBot\V12\Driver\Workerman\Worker;
 use OneBot\V12\Exception\OneBotFailureException;
-use OneBot\V12\MPUtils;
 use OneBot\V12\Object\Event\OneBotEvent;
 use OneBot\V12\RetCode;
 use Throwable;
